Improved shift_melody_low
Filling below the shifted content now is a better simulation of unobtrusive but likely noise, locally mimicking the noise in each time slice. Also detects when chunking behavior of the upstream time-pooling strategy leaves regions of amplitude -100, so fills with -100 below that region. The idea there is to not let that always so obviously expose to the NN where the filled region is.